router: change base router impl to a junit test class
any router impl will extend this test for a standard test suite
